requirements.txt
google-genai>=1.0.0
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.14
pytest>=8.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
